feat: extract components from send page
